more OSX warnings
- fix more OSX warnings

+++ hphp/runtime/base/runtime_option.cpp
@@ -477,7 +477,8 @@ static void setResourceLimit(int resource, Hdf rlimit, const char *nodeName) {
     }
     int ret = setrlimit(resource, &rl);
     if (ret) {
-      Logger::Error("Unable to set %s to %ld: %s (%d)", nodeName, rl.rlim_cur,
+      Logger::Error("Unable to set %s to %" PRId64 ": %s (%d)",
+                    nodeName, (int64_t)rl.rlim_cur,
                     Util::safe_strerror(errno).c_str(), errno);
     }
   }
